+++ utils/gmp_util.cpp
@@ -1,11 +1,12 @@
 #include <gmp.h>
 #include <gmpxx.h>
+#include <cstdint>
 #include "gmp_util.h"
 
 template <>
-mpz_class util::to_mpz(signed long long n)
+mpz_class util::to_mpz(int64_t n)
 {
-    long long n_abs = std::llabs(n);
+    int64_t n_abs = std::llabs(n);
     mpz_class result = (unsigned int)(n_abs >> 32);
     result <<= 32;
     result |= (unsigned int)(n_abs & 0xFFFFFFFF);
@@ -15,7 +16,7 @@ mpz_class util::to_mpz(signed long long n)
 }
 
 template <>
-mpz_class util::to_mpz(unsigned long long n)
+mpz_class util::to_mpz(uint64_t n)
 {
     mpz_class result = (unsigned int)((n >> 32) & 0xFFFFFFFF);
     result <<= 32;
